requirements.txt
python-telegram-bot
asyncpg
psycopg2-binary
langchain-google-genai